requirements.txt
streamlit
Pillow
torch==2.4.1
torchvision==0.19.1
